refactor : domain 파일 분리
