+++ src/views/AmountPaidInput.js
@@ -1,5 +1,5 @@
 import readLineAsync from '../utils/readLineAsync.js';
-import { buyNumbers } from '../domain/purchase.js';
+import { buyNumbers } from '../domain/amountPaid.js';
 
 const AmountPaidInput = async () => {
   const amountPaid = await readLineAsync('구입금액을 입력해 주세요');
